refactor: extract checkout components and modularize payment logic for improved maintainability
